requirements.txt
# Core framework
streamlit>=1.28.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# ONNX support
onnx>=1.14.0
onnx2pytorch>=0.4.1
onnxruntime>=1.15.0

# YOLO support
ultralytics>=8.0.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
